django/simple_chat/requirements.txt
django
django-df-chat==0.1.0

